some dependency removals and setup for refactor
